Computation_and_visualization_tool/Tensor_field_computation/requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.8.0
attrs==19.1.0
backports-abc==0.5
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
basemap==1.1.0
beautifulsoup4==4.3.2
bleach==3.0.2
certifi==2019.9.11
chardet==3.0.4
Cheetah==2.4.4
Click==7.0
cloudpickle==1.2.2
configparser==3.5.0
cryptography==2.1.4
cssselect==0.9.1
cycler==0.10.0
Cython==0.29.13
dap==2.2.6.7
dash==1.2.0
dash-core-components==1.1.2
dash-html-components==1.0.1
dash-renderer==1.0.1
dash-table==4.2.0
dask==1.1.5
decorator==4.4.1
defusedxml==0.5.0
dnspython==1.15.0
docopt==0.6.2
entrypoints==0.3
enum34==1.1.6
Flask==1.1.1
Flask-Compress==1.4.0
FormEncode==1.3.0
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.17.1
futures==3.3.0
gast==0.2.2
geographiclib==1.50
geopy==1.21.0
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.10.0
html5lib==0.9999999
httplib2==0.9.2
idna==2.8
imageio==2.6.1
imgaug==0.3.0
imutils==0.5.2
ipaddress==1.0.22
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==5.2.0
jupyter-core==4.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.4
mistune==0.8.4
mock==3.0.5
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
nltk==3.4.5
notebook==5.7.4
num2words==0.5.9
numpy==1.16.6
olefile==0.45.1
opencv-python==4.1.0.25
opencv-python-headless==4.1.2.30
pandas==0.24.2
pandocfilters==1.4.2
Paste==2.0.3
PasteDeploy==1.5.2
PasteScript==2.0.2
pathlib2==2.3.5
pbr==5.1.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
plotly==4.1.1
POT==0.6.0
prometheus-client==0.5.0
prompt-toolkit==1.0.18
protobuf==3.9.1
ptyprocess==0.6.0
pycrypto==2.6.1
Pygments==2.4.2
pygobject==3.26.1
pyOpenSSL==17.5.0
pyparsing==2.4.5
pyproj==1.9.5.1
pyrsistent==0.15.4
pyshp==1.2.12
pytesseract==0.2.7
python-dateutil==2.8.1
python-openid==2.2.5
python-utils==2.3.0
pytz==2019.3
PyWavelets==1.0.3
pyxdg==0.25
PyYAML==5.1.2
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.22.0
retrying==1.3.3
scandir==1.10.0
scgi==1.13
scikit-image==0.14.5
scikit-learn==0.20.3
scipy==1.2.2
seaborn==0.9.0
SecretStorage==2.3.1
Send2Trash==1.5.0
Shapely==1.6.4.post2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
subprocess32==3.5.4
Tempita==0.5.2
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==5.1.1
traitlets==4.3.3
urllib3==1.25.3
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.6
widgetsnbextension==3.4.2
wrapt==1.11.2
wxPython==3.0.2.0
wxPython-common==3.0.2.0
xlrd==1.2.0
